+++ packages/cli/src/zosjobs/submit/local-file/localFile.definition.ts
@@ -62,6 +62,24 @@ export const LocalFileDefinition: ICommandDefinition = {
             name: "jcl-symbols", aliases: ["js"],
             description: strings.COMMON_OPT.JCL_SYMBOLS,
             type: "string"
+        },
+        {
+            name: "job-encoding", aliases: ["je"],
+            description: strings.COMMON_OPT.JOB_ENCODING,
+            type: "string"
+        },
+        {
+            name: "job-record-length", aliases: ["jrl"],
+            description: strings.COMMON_OPT.JOB_RECORD_LENGTH,
+            type: "number",
+            defaultValue: 80
+        },
+        {
+            name: "job-record-format", aliases: ["jrf"],
+            description: strings.COMMON_OPT.JOB_RECORD_FORMAT,
+            type: "string",
+            allowableValues: {values: ["F", "V"]},
+            defaultValue: "F"
         }
     ]as ICommandOptionDefinition[]),
     profile: {